+++ docker/qgisserver/tests/functional/conftest.py
@@ -2,11 +2,10 @@
 
 import os
 import pytest
-from logging.config import fileConfig
+from unittest.mock import Mock, patch
 
-import plaster
-from sqlalchemy.exc import DBAPIError
 import transaction
+from qgis.server import QgsServerInterface
 
 from c2cgeoportal_commons.testing.initializedb import init_db, truncate_tables
 from c2cgeoportal_commons.testing import get_engine, get_session_factory, get_tm_session, generate_mappers
@@ -30,3 +29,50 @@ def settings():
     config.init(os.environ.get('GEOMAPFISH_CONFIG', '/etc/qgisserver/geomapfish.yaml'))
     settings.update(config.get_config())
     return settings
+
+
+@pytest.fixture(scope='module')
+def scoped_session(dbsession):
+    with patch('geomapfish_qgisserver.accesscontrol.scoped_session',
+               return_value=dbsession) as mock:
+        yield mock
+
+
+@pytest.fixture(scope='function')
+def server_iface():
+    yield Mock(spec=QgsServerInterface)
+
+
+@pytest.fixture(scope='module')
+def qgs_access_control_filter():
+    """
+    Mock some QgsAccessControlFilter methods:
+    - __init__ which does not accept a mocked QgsServerInterface;
+    - serverInterface to return the right server_iface.
+    """
+    class DummyQgsAccessControlFilter():
+
+        def __init__(self, server_iface):
+            self.server_iface = server_iface
+
+        def serverInterface(self):
+            return self.server_iface
+
+    with patch.multiple('geomapfish_qgisserver.accesscontrol.QgsAccessControlFilter',
+                        __init__=DummyQgsAccessControlFilter.__init__,
+                        serverInterface=DummyQgsAccessControlFilter.serverInterface) as mocks:
+        yield mocks
+
+
+@pytest.fixture(scope='class')
+def single_ogc_server_env(dbsession):
+    os.environ['GEOMAPFISH_OGCSERVER'] = 'qgisserver1'
+    yield
+    del os.environ['GEOMAPFISH_OGCSERVER']
+
+
+@pytest.fixture(scope='class')
+def multiple_ogc_server_env(dbsession):
+    os.environ['GEOMAPFISH_ACCESSCONTROL_CONFIG'] = '/etc/qgisserver/multiple_ogc_server.yaml'
+    yield
+    del os.environ['GEOMAPFISH_ACCESSCONTROL_CONFIG']
